refactor: Reorganize src directory into logical module structure
Reorganize 12 flat modules into 3 logical directories:

- capture/: Screen capture domain
  - capture_loop.rs (renamed from capture.rs)
  - image_store.rs, metadata.rs, ocr.rs

- data/: Data persistence and reporting
  - database.rs, report.rs

- infra/: Infrastructure concerns
  - config.rs, error.rs, logging.rs, pause_control.rs

Each directory has a mod.rs with appropriate re-exports.
All 50 tests pass, CLI functionality preserved.
